Run e2e tests on alfajores (#258)
Modifies our e2e tests to be able to run on alfajores if the
environment variable `NETWORK` is set to 'alfajores'.
Also adds a CI step to run them against alfajores.

Note that the `test_fee_currency...` tests are skipped
because they require the ability to deploy new fee currencies,
which we probably don't want to be doing on a live network
as it could cause issues for people, and we also don't want to
expose the credentials to do so.

+++ e2e_test/run_all_tests.sh
@@ -6,32 +6,45 @@ source "$SCRIPT_DIR/shared.sh"
 
 TEST_GLOB=$1
 
-## Start geth
-cd "$SCRIPT_DIR/.." || exit 1
-make geth
-trap 'kill %%' EXIT # kill bg job at exit
-build/bin/geth --dev --http --http.api eth,web3,net --txpool.nolocals &>"$SCRIPT_DIR/geth.log" &
+if [ -z $NETWORK ]; then
+    ## Start geth
+    cd "$SCRIPT_DIR/.." || exit 1
+    make geth
+    trap 'kill %%' EXIT # kill bg job at exit
+    build/bin/geth --dev --http --http.api eth,web3,net --txpool.nolocals &>"$SCRIPT_DIR/geth.log" &
+    
+    # Wait for geth to be ready
+    for _ in {1..10}; do
+    	if cast block &>/dev/null; then
+    		break
+    	fi
+    	sleep 0.2
+    done
+    
+    ## Run tests
+    echo Geth ready, start tests
+fi
 
-# Wait for geth to be ready
-for _ in {1..10}; do
-	if cast block &>/dev/null; then
-		break
-	fi
-	sleep 0.2
-done
-
-## Run tests
-echo Geth ready, start tests
 cd "$SCRIPT_DIR" || exit 1
-
 # There's a problem with geth return errors on the first transaction sent.
 # See https://github.com/ethereum/web3.py/issues/3212
 # To work around this, send a transaction before running tests
-cast send --async --json --private-key "$ACC_PRIVKEY" "$TOKEN_ADDR" 'transfer(address to, uint256 value) returns (bool)' 0x000000000000000000000000000000000000dEaD 100
+cast send --json --private-key "$ACC_PRIVKEY" "$TOKEN_ADDR" 'transfer(address to, uint256 value) returns (bool)' 0x000000000000000000000000000000000000dEaD 100 > /dev/null || true
 
 failures=0
 tests=0
+echo "Globbing with \"$TEST_GLOB\""
 for f in test_*"$TEST_GLOB"*; do
+	echo "for file $f"
+	if [[ -n $NETWORK ]]; then
+		case $f in
+		  # Skip tests that require a local network.
+		  test_fee_currency_fails_on_credit.sh|test_fee_currency_fails_on_debit.sh|test_fee_currency_fails_intrinsic.sh)
+		  echo "skipping file $f"
+		  continue
+		  ;;
+	    esac
+	fi
 	echo -e "\nRun $f"
 	if "./$f"; then
 		tput setaf 2 || true
